Parking sensor implementation
By merging the LCD and ultrasonic sensor, we will create an affordable, user-friendly parking sensor. When the object is far away, the LCD should show green; when it is in, it should show yellow; and when it is too close, it should show red. The state it was in prior to turning yellow determines the yellow's boundaries. Additionally, by going into the end state, we will include turning off the parking sensor.